% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_code.R
\name{prep_code}
\alias{prep_code}
\title{Create good enuff code files}
\usage{
prep_code(folder_name = NULL, header = TRUE)
}
\arguments{
\item{folder_name}{name of folder for code files (string)}

\item{header}{include roxygen2 headers?}
}
\description{
Create the code folder for the import, tidy, wrangle, visualize,
    model, and data .R files.
}
\details{
If header is FALSE, code files are created with a standard script
   header. By default, the code files are placed in a code/ folder, but this
   can be changed with a character string.
}
