% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_hdr.R
\name{code_hdr}
\alias{code_hdr}
\title{Insert good enuff R script header}
\usage{
code_hdr()
}
\description{
The code_hdr() function can be used in a new R script
    to create a header. The header includes sections for:
- what the code creates
- who the author is
- MIT License
- Version
}
\examples{
# use this to create a script header (with RStudio open)
# code_hdr("import")

}
