% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_folders.R
\name{prep_folders}
\alias{prep_folders}
\title{good enuff R setup}
\usage{
prep_folders(code = TRUE, data = TRUE, docs = TRUE)
}
\arguments{
\item{code}{logical, include results code and files?}

\item{data}{logical, include data folder and files?}

\item{docs}{logical, include docs folder and files?}
}
\description{
prep_folders() is a wrapper around the prep_code(), prep_data(),
    and prep_docs() functions. This is intended for an easy setup with a
    new project.
}
